+++ Data/ByteString/Base64/URL/Lazy.hs
@@ -20,6 +20,8 @@ module Data.ByteString.Base64.URL.Lazy
       encode
     , encodeUnpadded
     , decode
+    , decodeUnpadded
+    , decodePadded
     , decodeLenient
     ) where
 
@@ -55,6 +57,22 @@ decode b = -- Returning an Either type means that the entire result will
            case B64.decode $ S.concat $ L.toChunks b of
            Left err -> Left err
            Right b' -> Right $ L.fromChunks [b']
+
+-- | Decode a unpadded base64url-encoded string, failing if input is padded.
+-- This function follows the specification in <http://tools.ietf.org/rfc/rfc4648 RFC 4648>
+-- and in <https://tools.ietf.org/html/rfc7049#section-2.4.4.2 RFC 7049 2.4>
+decodeUnpadded :: L.ByteString -> Either String L.ByteString
+decodeUnpadded bs = case B64.decodeUnpadded $ S.concat $ L.toChunks bs of
+  Right b -> Right $ L.fromChunks [b]
+  Left e -> Left e
+
+-- | Decode a padded base64url-encoded string, failing if input is improperly padded.
+-- This function follows the specification in <http://tools.ietf.org/rfc/rfc4648 RFC 4648>
+-- and in <https://tools.ietf.org/html/rfc7049#section-2.4.4.2 RFC 7049 2.4>
+decodePadded :: L.ByteString -> Either String L.ByteString
+decodePadded bs = case B64.decodePadded $ S.concat $ L.toChunks bs of
+  Right b -> Right $ L.fromChunks [b]
+  Left e -> Left e
 
 -- | Decode a base64-encoded string.  This function is lenient in
 -- following the specification from